Services component is refactored
