+++ src/components/Services/index.js
@@ -1,13 +1,17 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 
 import { useNavigate } from '@reach/router';
 import { Grid } from '@material-ui/core';
+
+import { ListOfWords, selectedCategory } from '../../utils';
+
 import HeadOurServices from './HeadOurServices';
 import ItemService from './ItemService';
 import ViewerService from './ViewerService';
 
 const Services = ({
   services,
+  category,
   selected,
   title,
   subtitle,
@@ -15,6 +19,7 @@ const Services = ({
   colorTitle,
 }) => {
   const [selectedIndex, setSelectedIndex] = useState(0);
+  const [selectedCategories, setSelectedCategories] = useState([]);
 
   const navigate = useNavigate();
 
@@ -22,16 +27,14 @@ const Services = ({
     const newIndex = services.findIndex(
       (v) => v.node.title.toLowerCase() === s.toLowerCase(),
     );
+
+    setSelectedCategories(
+      selectedCategory(services[newIndex], category),
+    );
     if (newIndex >= 0) {
       setSelectedIndex(newIndex);
     }
   };
-
-  useEffect(() => {
-    if (selected) {
-      selectService(selected);
-    }
-  });
 
   return (
     <Grid item xs={12} container justify="center">
@@ -67,7 +70,11 @@ const Services = ({
           ))}
       </Grid>
       {services.length > 0 ? (
-        <ViewerService service={services[selectedIndex]} />
+        <ViewerService
+          service={services[selectedIndex]}
+          words={ListOfWords(services)}
+          categories={selectedCategories}
+        />
       ) : null}
     </Grid>
   );
